Write the `done` file on next tick.
In automatic mode, a file is written to indicate that the screenshot has
been completed. That enables the CLI to shutdown Factorio. However, if
that file is written in the same tick as the screenshot, the ordering
might not be guaranteed - the file might be visible quickly, leading to
Factorio to be shutdown before the other files are completely written.

It seems that by waiting one tick before writing that file, it is enough
to have the screenshot written. On Linux, this is still sending an
interrupt signal to shutdown Factorio cleanly, but that allows to use a
more direct kill signal Windows (which lacks interrupt signal).