Do not use frontend code from the mod when serving from CLI.
The mod generated static index.html and javascript files with each
mapshot. Those are useful for serving the whole content statically.

However, when using the CLI (through `dev` or `serve`), instead use the
version in the CLI (or built from NPM). This allows to have a fresh
version and makes development easier.