Make `render` ignore Factorio return code when render is done.
I haven't found a good way to have a useful error code from Factorio
when shutting it down from `render` - so let's just ignore the error
code from Factorio as long as the render was indicated as `done` (which
should be good enough in the context of the `render` command).

This does not change the `dev` subcommand - it has a different usecase,
and does not have a clear-cut "done" condition.

fixes #28