Make `surface` setting functional.
A setting was added a while ago to, in theory, control which surfaces to
render. Unfortunately, the setting was a no-op, as the code to do the
filtering was just not there.

This change fixes that - it is now possible to select one (or more)
surface to render. Still defaults to render all surfaces.

Thanks to @rdbruijne for providing the fix.

Fixes #30