Be mindful of filepath encoding.
Factorio save are limited in the character set they accept, but still -
things like spaces and square brackets are allowed.

As Go mux expects very specific forms of encoding, we need to be careful
of what we're handling around. Specifically, it is easy to have requests
from the browser end up having spaces encoded in the paths, but not the
square bracket. This in turns fail, as Go does not find the
corresponding path.

Instead, calculate the expected URL path on the Go side and send that
for the frontend to use as-is. This also requires the frontend to be a
bit careful with it, to not accidentely decode it.

I'm not super confident here - it fixes things in my tests and kinda
make sense. However, I'm a bit puzzled a Go behavior here with partial
encoding and I wonder if I missed something.

Fixes #51