Fix generation of mapshot config for static index.html.
In recent encoding fixes, the viewer code has been changed to expect an
`encoded_path` instead of just a `path`. Unfortunately, I missed the
case of the static index.html generated with each shot, which was still
refering to just `path`.

This commit fixes that. Luckily, this only needs to encode the path to
the hash directory containing the actual generated data - and thus I can
away without having to do URL encoding in Lua, as this generated path
name does not contain any special character.

Fixes #56