Deactivate clouds when rendering from CLI.
When zoomed out, clouds make for weird artifact, so deactivate them.
Unfortunately, that only works when rendering from the CLI. I've tried
doing the same when using the `/mapshot` in-game command, but I have not
found away to restore the clouds settings in that case. Even waiting for
the end of screenshotting & a tick does not seem to be enough.

Issue #50